Try average cluster, add delta mfcc

+++ process_rawdata.py
@@ -71,7 +71,7 @@ def breakDown(wave_data, fs, dovisualize=False):
             bp += [i]
             i += interval
         elif(energy[i] < threshold[i]):
-            min_i = findMin(energy, i-interval/2, i+interval/2)
+            min_i = findMin(energy, int(i-interval/2), int(i+interval/2))
             if(min_i == i):
                 bp += [i]
                 i += interval
